# src/propostas_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "3RN Crédito Popular API",
        "description": "API para gestão de propostas de crédito consignado",
        "contact": {
            "email": "suporte@3rncredito.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://propostas_app.localhost/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/propostas": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Listar todas as propostas",
                "description": "Retorna uma lista de propostas com filtros opcionais por status e nome/CPF.",
                "operationId": "b8c8ece26a54c07823ea4fa515ebd8ae",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nome ou CPF",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Maria Silva"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "rascunho",
                                "em_analise",
                                "aprovada",
                                "reprovada",
                                "cancelada"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de propostas retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Propostas carregadas com sucesso."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "cliente_nome": {
                                                        "type": "string",
                                                        "example": "Maria da Silva"
                                                    },
                                                    "cliente_cpf": {
                                                        "type": "string",
                                                        "example": "12345678900"
                                                    },
                                                    "cliente_salario": {
                                                        "type": "number",
                                                        "example": 5000
                                                    },
                                                    "valor_solicitado": {
                                                        "type": "number",
                                                        "example": 2000
                                                    },
                                                    "prazo_meses": {
                                                        "type": "integer",
                                                        "example": 24
                                                    },
                                                    "taxa_juros": {
                                                        "type": "number",
                                                        "example": 2.5
                                                    },
                                                    "valor_parcela": {
                                                        "type": "number",
                                                        "example": 150
                                                    },
                                                    "valor_total": {
                                                        "type": "number",
                                                        "example": 3600
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "em_analise"
                                                    },
                                                    "observacoes": {
                                                        "type": "string",
                                                        "example": "Cliente com bom histórico de crédito."
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-11-08 14:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-11-08 14:30:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}